# Use strings.Index instead of strings.Split