@cpuguy83 2020-04-24
btw, here is the diff without changing how we allocate the original slice (so just changing strings.Split):

```
benchmark                                     old allocs     new allocs     delta
BenchmarkReplaceOrAppendEnvValues/0-8         5              1              -80.00%
BenchmarkReplaceOrAppendEnvValues/100-8       110            6              -94.55%
BenchmarkReplaceOrAppendEnvValues/1000-8      1013           9              -99.11%
BenchmarkReplaceOrAppendEnvValues/10000-8     10022          18             -99.82%

benchmark                                     old bytes     new bytes     delta
BenchmarkReplaceOrAppendEnvValues/0-8         192           64            -66.67%
BenchmarkReplaceOrAppendEnvValues/100-8       7360          4032          -45.22%
BenchmarkReplaceOrAppendEnvValues/1000-8      64832         32704         -49.56%
BenchmarkReplaceOrAppendEnvValues/10000-8     1146049       825920        -27.93%
```